+++ processor/src/test/resources/expected/com/example/arez/ComponentWithCallbackCustomeParamNames_.java
@@ -1,0 +1,15 @@
+package com.example.arez;
+
+import javax.annotation.Generated;
+import javax.annotation.Nonnull;
+
+@Generated("react4j.processor.ReactProcessor")
+class ComponentWithCallbackCustomeParamNames_ {
+  private ComponentWithCallbackCustomeParamNames_() {
+  }
+
+  @Nonnull
+  static ComponentWithCallbackCustomeParamNames.CustomHandler _handleFoo2(@Nonnull final ComponentWithCallbackCustomeParamNames component) {
+    return ((React4j_ComponentWithCallbackCustomeParamNames) component)._handleFoo2;
+  }
+}